create custom hook for useClickOutside and useMatchMedia
