+++ frontend/src/hooks/useMatchMedia.js
@@ -1,0 +1,21 @@
+import { useEffect, useState } from 'react';
+
+export const useMatchMedia = (query) => {
+  const [matches, setMatches] = useState(window.matchMedia(query).matches);
+
+  useEffect(() => {
+    const mediaQuery = window.matchMedia(query);
+
+    const handleMediaQueryChange = (e) => {
+      setMatches(e.matches);
+    };
+
+    mediaQuery.addEventListener('change', handleMediaQueryChange);
+
+    return () => {
+      mediaQuery.removeEventListener('change', handleMediaQueryChange);
+    };
+  }, [query]);
+
+  return matches;
+};
